runtime_init_clocks() restores perfectly
No more consumption discrepancy.

Still do not understand what did I miss the first time?

+++ install.sh
@@ -1,13 +1,13 @@
 git clone https://github.com/raspberrypi/pico-sdk.git
-git clone https://github.com/raspberrypi/pico-extras.git
-cd pico-sdk
-git submodule update --init --recursive
-cd ..
 cp pico-sdk/external/pico_sdk_import.cmake .
-cp pico-extras/external/pico_extras_import.cmake .
 mkdir build
 cat << EOF > build/gregmake.sh
 make
+# Check if 'make' was successful
+if [ $? -ne 0 ]; then
+  echo "Build failed. Exiting."
+  exit 1
+fi
 picotool load test.uf2
 picotool reboot
 EOF
